% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{clean_data}
\alias{clean_data}
\title{Cleaned OWID COVID-19 dataset}
\format{
A data frame with columns: date, location, total_cases, new_cases, total_deaths, etc.
}
\source{
Data originally from \url{https://ourworldindata.org/}, cleaned for use in this package.
}
\description{
This dataset contains cleaned COVID-19 case data for selected countries between 2020 and 2023.
Filtered and preprocessed from the original OWID dataset.
}
\examples{
data(clean_data)
str(clean_data)
summary(clean_data)
}

