% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_project.R
\name{new_project}
\alias{new_project}
\title{Create New Project Scaffolding}
\description{
Initializes a new R project structure with optional settings such as package setup, GitHub integration, CI, and testing.
Useful for bootstrapping research or package development projects.
}
\usage{
new_project(
  name,
  path = ".",
  package = TRUE,
  github = FALSE,
  private.repo = TRUE,
  ci = "none",
  makefile = TRUE,
  pipe = TRUE,
  testthat = FALSE,
  verbose = FALSE,
  open.project = TRUE
)
}
\arguments{
  \item{name}{The name of the new project (required).}
  \item{path}{Directory where the project should be created. Defaults to current working directory.}
  \item{package}{Logical. Should the project be structured as an R package? Default is \code{TRUE}.}
  \item{github}{Logical. Should a GitHub repository be initialized? Default is \code{FALSE}.}
  \item{private.repo}{Logical. If creating a GitHub repository, should it be private? Default is \code{TRUE}.}
  \item{ci}{Continuous integration tool to configure. Options: \code{"none"}, \code{"github"}, \code{"travis"}.}
  \item{makefile}{Logical. Should a Makefile be included? Default is \code{TRUE}.}
  \item{pipe}{Logical. Should the pipe operator (\code{\%>\%}) be activated (e.g., via \code{magrittr} or \code{tidyverse})?}
  \item{testthat}{Logical. Should \code{testthat} structure be included for testing? Default is \code{FALSE}.}
  \item{verbose}{Logical. Print progress messages? Default is \code{FALSE}.}
  \item{open.project}{Logical. Open the new project in RStudio after creation? Default is \code{TRUE}.}
}
\examples{
new_project("covid19-analysis", github = TRUE, testthat = TRUE)
}
\seealso{
\url{https://github.com/yagizemreozbilge/R-template.git}
}
\author{
Yagiz Emre OZBILGE
}

